Add files via upload
Blackjack Game
Welcome to the Blackjack Game repository! This project is a simple yet engaging implementation of the classic casino card game Blackjack. The game is designed for a single player to play against the dealer.
Blackjack, also known as 21, is a popular card game played in casinos worldwide. The objective of the game is to beat the dealer by having a hand value closer to 21 without exceeding it.